Moved storm point details to its own component removed many unused imports from map.js added error cone component for ECCC data
